refactor: change blog location to /news
