+++ config.json
@@ -1,6 +1,6 @@
 {
-  "base_url": "https://sdohplace.org/blog",
-  "site_title": "Blog for SDOH & Place Project",
+  "base_url": "https://sdohplace.org/news",
+  "site_title": "News | SDOH & Place Project",
   "site_description": "News and updates related to the SDOH & Place Project from Heathy Regions & Policies Lab at Univiersity of Illinois, Urbana-Champaign.",
   "site_keywords": [
     {
