+++ src/components/news/Copyright.tsx
