+++ src/components/news/meta/OpenGraphMeta.tsx
