+++ src/pages/news/index.tsx
@@ -1,9 +1,9 @@
 import { GetStaticProps } from "next";
-import Layout from "@/components/blog/Layout";
-import BasicMeta from "@/components/blog/meta/BasicMeta";
-import OpenGraphMeta from "@/components/blog/meta/OpenGraphMeta";
-import TwitterCardMeta from "@/components/blog/meta/TwitterCardMeta";
-import PostList from "@/components/blog/PostList";
+import Layout from "@/components/news/Layout";
+import BasicMeta from "@/components/news/meta/BasicMeta";
+import OpenGraphMeta from "@/components/news/meta/OpenGraphMeta";
+import TwitterCardMeta from "@/components/news/meta/TwitterCardMeta";
+import PostList from "@/components/news/PostList";
 import config from "../../lib/config";
 import { countPosts, listPostContent, PostContent } from "../../lib/posts";
 import { listTags, TagContent } from "../../lib/tags";
@@ -18,7 +18,7 @@ type Props = {
   };
 };
 export default function Index({ posts, tags, pagination }: Props) {
-  const url = "/blog";
+  const url = "/news";
   const title = "All posts";
   return (
     <Layout>
